fix: logic related expect/actual classes pattern
