+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/network/NetworkStateObserverImpl.kt
@@ -23,9 +23,9 @@ import android.net.ConnectivityManager
 import android.net.Network
 import android.net.NetworkCapabilities
 import android.os.Build
-import com.wire.kalium.logger.KaliumLogLevel
 import com.wire.kalium.common.logger.kaliumLogger
 import com.wire.kalium.common.logger.logStructuredJson
+import com.wire.kalium.logger.KaliumLogLevel
 import com.wire.kalium.network.NetworkState
 import com.wire.kalium.network.NetworkStateObserver
 import com.wire.kalium.util.KaliumDispatcher
@@ -165,7 +165,7 @@ internal actual class NetworkStateObserverImpl(
         }
     }
 
-    override fun observeNetworkState(): StateFlow<NetworkState> = networkStateFlow
+    actual override fun observeNetworkState(): StateFlow<NetworkState> = networkStateFlow
 
     private sealed class DefaultNetworkData {
         data object NotConnected : DefaultNetworkData()
